+++ app/src/main/res/values-es-rES/strings.xml
@@ -11,7 +11,6 @@
     <string name="main_menu">Menu</string>
     <string name="new_game">EMPEZAR</string>
     <string name="btnAcabar">Terminar</string>
-    <string name="texto10_1_a4">• No</string>
     <string name="enunciadoCuento">" Necesitas tomar una foto de la granja y escribir una pequeña historia."</string>
     <string name="texto0_a2">Varios secretos del pasado alrededor de Andra Mari de Ermu han comenzado a revelarse a través de la arqueología. Los arqueólogos han estado trabajando en dos sitios, uno de 12 m ^ 2 y otro de 20 m ^ 2, bajo el coro de la iglesia de Andra Mari de Ermua y al lado del campanario. Las excavaciones han revelado una necrópolis, construida en el siglo XI. y XII. con entierros en los siglos XVI y XVII. Otra parcela de la iglesia del siglo XVI. De esto podemos deducir que el XII. que demolieron el edificio del siglo XIX para construir lo que es hoy. Por otra parte, XII. se han encontrado monedas del siglo, riquezas asociadas con entierros. Esto nos lleva a pensar que estaba asociado con la oligarquía.</string>
     <string name="texto0_a4">La historia de la Bruja de Lezeaga se ha escuchado durante mucho tiempo en todas partes de Llodio. Hace mucho tiempo, una niña del pueblo tenía que pasar por Sorginzulo en busca de agua todos los días. Cansado de hacer siempre el mismo giro, dijo: \"¿No pueden las brujas construir un puente a cambio de que me una a ellos?\" A la mañana siguiente, los aldeanos se despertaron y encontraron un nuevo puente y esta pequeña niña se convirtió en bruja. Desde ese día, vivió en la cueva de Lezeaga. Permaneció para siempre excluido de sus compatriotas. A veces, con su hermosa voz, atraía al pastor mientras se peinaba con un cepillo dorado.</string>
@@ -47,4 +46,8 @@
     <string name="texto3_2_a2">Restos.</string>
     <string name="texto2_a3">Hace mucho tiempo, años y años atrás, en una noche como la noche anterior, los pastores decidieron llevar las ovejas a Urkiola. Santa Apolonia, consciente de esto, decidió ir con ellos a pasar la noche; por lo tanto, él ayudaría y protegería esto. Santa Apolonia, aunque protectora de los dientes, no era del todo hábil, y al atravesar esta ermita pisó una piedra y dejó su huella. Desde entonces, según la tradición, puede tomar agua de la fuente mágica en Santa Lucía y pisar esta huella tres veces para garantizar la salud bucal durante un año.</string>
     <string name="texto1_a7">Dolumin se celebra el viernes antes de Pascua. En esta celebración, se celebra una feria de animales desde 1980 y reúne a personas de la zona de Llodio. En el pasado, la feria se basaba en el comercio de ganado, pero hoy se centra en la exhibición de animales exóticos y animales. Durante este día, se realizan deportes recreativos, es decir, deportes populares y actividades religiosas. Es uno de los festivales más importantes de la región y es un festival que recuerda la memoria de la población rural de Llodio.</string>
+    <string name="texto1_p2">Autores</string>
+    <string name="texto2_p2">Galeria</string>
+    <string name="texto3_p2">Mapa</string>
+    <string name="texto4_p2">Patrimonios</string>
 </resources>
